refactor: rename aimbat._types to aimbat.types
