--- src/aimbat/_types/__init__.py
@@ -1,29 +1,0 @@
-# flake8: noqa: E402, F403
-"""Custom types used in AIMBAT.
-
-Three modules are re-exported from here:
-
-- `_enums` — `EventParameter` and `SeismogramParameter` StrEnums for CLI arg
-  typing and validation.
-- `_pydantic` — `PydanticTimestamp`, `PydanticTimedelta`, and constrained
-  variants (`PydanticPositiveTimedelta`, `PydanticNegativeTimedelta`,
-  `PydanticNonNegativeTimedelta`, `PydanticNonNegativeFloat`) for use in
-  Pydantic models.
-- `_sqlalchemy` — `SAPandasTimestamp` and `SAPandasTimedelta` SQLAlchemy
-  `TypeDecorator` classes for storing `pandas.Timestamp` and
-  `pandas.Timedelta` values in the database.
-"""
-
-from .._utils import export_module_names
-
-_internal_names = set(dir())
-
-from ._enums import *
-from ._pydantic import *
-from ._sqlalchemy import *
-
-__all__ = [s for s in dir() if not s.startswith("_") and s not in _internal_names]
-
-export_module_names(globals(), __name__)
-
-del _internal_names
