+++ src/aimbat/types/_enums.py
