+++ src/aimbat/types/_pydantic.py
