+++ zensical.toml
@@ -31,7 +31,6 @@ nav = [
     { "MCCC Alignment" = "usage/mccc.md" },
     { "Quality Assessment" = "usage/quality.md" },
     { "Exporting Results" = "usage/results.md" },
-    { "Exporting Results" = "usage/results.md" },
     { "API Basics" = "usage/api-basics.md" },
     { "Aimbat Defaults" = "usage/defaults.md" },
     { "Recipes" = "usage/recipes.md" },
@@ -46,10 +45,10 @@ nav = [
     "api/aimbat/models.md",
     "api/aimbat/plot.md",
     "api/aimbat/utils.md",
+    "api/aimbat/types.md",
     "api/aimbat/_cli.md",
     "api/aimbat/_tui.md",
-    "api/aimbat/_types.md",
-    # "api/aimbat/ui.md",
+    # "api/aimbat/_ui.md",
   ] },
   { "Development" = [
     { "Database Models" = "development/database-models.md" },
